crewai_mcp_course/requirements.txt
crewai>=0.80.0
mcp>=1.0.0
fastmcp>=2.0.0
python-dotenv>=1.0.0
